% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promise.R
\name{asyncReadLines}
\alias{asyncReadLines}
\title{A promise-based text file reader}
\usage{
asyncReadLines(file_path)
}
\arguments{
\item{file_path}{The path to the file to read}
}
\value{
A promise that resolves to an array of lines from the file
}
\description{
This function reads a text file asynchronously
on another thread and returns a promise that resolves to an array
of lines from the file.
}
