% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asyncWriteLines.R
\name{asyncWriteLines}
\alias{asyncWriteLines}
\title{A promise-based text file writer}
\usage{
asyncWriteLines(content, path, append = FALSE)
}
\arguments{
\item{content}{A vector of lines to write to the file}

\item{path}{The path to the file to write}

\item{append}{A logical value indicating whether to append to the file}
}
\value{
A promise that resolves to the number of bytes written
}
\description{
This function writes a vector of lines
to a text file asynchronously on another thread
and returns a promise that resolves to the number of bytes written
}
